cleanArch
