--- internal/app/httpserver/hellohandler.go
@@ -1,10 +1,0 @@
-package httpserver
-
-import (
-	"io"
-	"net/http"
-)
-
-func HelloHandler(w http.ResponseWriter, req *http.Request) {
-	io.WriteString(w, "hello, world!\n")
-}
