+++ internal/models/email_model.go
@@ -1,0 +1,12 @@
+package models
+
+import "time"
+
+type Email struct {
+	Author      string    `json:"author"`
+	Description string    `json:"description"`
+	Text        string    `json:"text"`
+	Badge_text  string    `json:"badge_text"`
+	Badge_type  string    `json:"badge_type"`
+	Date        time.Time `json:"date"`
+}
