+++ internal/models/error_model.go
@@ -1,0 +1,6 @@
+package models
+
+type Error struct {
+	Status int    `json:"status"`
+	Body   string `json:"body"`
+}
