+++ internal/models/session_model.go
@@ -1,0 +1,12 @@
+package models
+
+import (
+	"time"
+)
+
+type Session struct {
+	Name      string
+	ID        string
+	Time      time.Time
+	UserLogin string
+}
